scripts/tinkering/hr_poller_requirements.txt
paho-mqtt<2.0.0

bitstruct
bleak
